+++ test-argv.js
@@ -1,0 +1,15 @@
+console.log('process.argv:', process.argv);
+console.log('argv[0]:', process.argv[0]);
+console.log('argv[1]:', process.argv[1]);
+console.log('__filename:', __filename);
+console.log('__dirname:', __dirname);
+
+// Try to get main module
+if (require.main) {
+  console.log('require.main.filename:', require.main.filename);
+}
+
+// Check import.meta
+if (import.meta && import.meta.url) {
+  console.log('import.meta.url:', import.meta.url);
+}